use usize for max youtube count

+++ src/post.rs
@@ -8,7 +8,7 @@ use uuid::Uuid;
 const APPLICATION_OCTET_STREAM: &'static str = "application/octet-stream";
 const UPLOADS_DIR: &'static str = "uploads";
 const MEDIA_DIR: &'static str = "pub/media";
-const MAX_YOUTUBE_EMBEDS: i32 = 3;
+const MAX_YOUTUBE_EMBEDS: usize = 3;
 
 #[derive(sqlx::Type, serde::Serialize, serde::Deserialize, PartialEq, Clone, Debug)]
 #[serde(rename_all = "snake_case")]
